requirements.txt
PyYAML>=6.0.1
akshare>=1.11.95
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.20.0
streamlit>=1.34.0
requests>=2.31.0